requirements.txt
imageio==2.5.0
marshmallow==3.0.5
numpy==1.17.2
Pillow==6.1.0
